% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{apc}
\alias{apc}
\alias{apc.surveil}
\alias{apc.stand_surveil}
\title{Annual and cumulative percent change}
\usage{
apc(x)

\method{apc}{surveil}(x)

\method{apc}{stand_surveil}(x)
}
\arguments{
\item{x}{For the \code{apc} function, \code{x} may be a fitted \code{surviel} model, or standardized rates (a \code{stand_surveil} object). For print and plot methods, \code{x} is the \code{apc_ls} object returned by \code{apc}.}
}
\value{
An \code{apc_ls} (list) object containing two data frames:
\describe{
\item{apc}{A data frame containing a summary of the posterior distribution for period-specific percent change. This contains the posterior mean (\code{apc}) 95 percent credible intervals (\code{lwr} and \code{upr} bounds).}
\item{cpc}{A data frame containing a summary of the posterior distribution for the cumulative percent change in risk at each time period. This contains the posterior mean (\code{cpc}) and 95 percent credible interval (\code{lwr} and \code{upr} bounds).}
}
}
\description{
Summarize annual and cumulative percent change in risk
}
\examples{
data(cancer)
\dontrun{
 fit <- stan_rw(cancer, time = Year, group = Age)
 x <- apc(fit)
 print(x, rate = 10e3)
 plot(x, cumulative = TRUE)
}
}
\seealso{
\code{\link[surveil]{stan_rw}} \code{\link[surveil]{standardize}}
}
