% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{print.apc_ls}
\alias{print.apc_ls}
\alias{plot.apc_ls}
\title{Methods for APC objects}
\usage{
\method{print}{apc_ls}(x, digits = 0, max = 10, ...)

\method{plot}{apc_ls}(
  x,
  cumulative = FALSE,
  col = "black",
  fill = "black",
  alpha = 0.5,
  base_size = 14,
  lwd = 1,
  ...
)
}
\arguments{
\item{x}{An \code{apc_ls} object returned by \code{\link[surveil]{apc}}}

\item{digits}{Print this many digits (passed to \code{\link[base]{print.data.frame}})}

\item{max}{Maximum number of time periods (rows) to print}

\item{...}{Additional arguments}

\item{cumulative}{Plot cumulative percent change? Defaults to \code{cumulative = FALSE}}

\item{col}{Line color}

\item{fill}{Fill color for the 95 percent credible interval}

\item{alpha}{Transparency (defaults to \code{alpha = 0.5}) for the credible interval (passed to \code{\link[ggplot2]{geom_ribbon}}).}

\item{base_size}{Size of plot attributes, passed to `\code{\link[ggplot2]{theme_classic}}}

\item{lwd}{Line width}
}
\description{
Methods for APC objects
}
\details{
\subsection{print.apc_ls}{

Any additional arguments (\code{...}) to the print method will be  passed to \code{\link[base]{print.data.frame}}
}

\subsection{plot.apc_ls}{

Any additional arguments (\code{...}) will be passed to \code{\link[ggplot2]{theme}}.
}
}
