% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{group_diff}
\alias{group_diff}
\title{Calculate summary measures of inequality: rate ratios, rate differences, excess cases}
\usage{
group_diff(fit, target, reference, samples = FALSE)
}
\arguments{
\item{fit}{A fitted \code{surveil} time series model}

\item{target}{The name (character string) of the disadvantaged group that is the target of inference.}

\item{reference}{The name (character string) of the reference group to which \code{target} will be compared.}

\item{samples}{If \code{TRUE}, return MCMC samples for the inequality measures. The default is \code{FALSE}.}
}
\value{
A list, also of class "surveil_diff", with the following elements ...
}
\description{
Calculate pairwise measures of health inequality from a fitted \code{surveil} time series model, with credible intervals and MCMC samples.
}
\author{
Connor Donegan (Connor.Donegan@UTSouthwestern.edu)
}
