% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.theil}
\alias{plot.theil}
\title{Plot theil's index}
\usage{
\method{plot}{theil}(x, col = "black", fill = "gray80", scale = 100, labels = x$summary$time)
}
\arguments{
\item{x}{An object of class \code{theil}, as created by calling \code{\link[surveil]{theil}} on a fitted \code{surveil} model.}

\item{fill}{Fill and line color}

\item{scale}{Multiply Theil's T by \code{scale} for readability}

\item{labels}{X-axis labels (time periods)}
}
\value{
A \code{ggplot}
}
\description{
Plot theil's index
}
\examples{
 library(tidyverse)
 dfw <- filter(msa, str_detect(MSA, "Dallas")
 fit <- stan_rw(dfw, group = Race, iter = 500)
 theil.dfw <- theil(fit)
 plot(theil.dfw)
}
