% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.theil_list}
\alias{plot.theil_list}
\title{Plot theil's index for nested data}
\usage{
\method{plot}{theil_list}(
  x,
  col = "black",
  fill = "black",
  alpha = 0.25,
  between_title = "Between",
  within_title = "Within",
  total_title = "Total",
  scale = 100,
  plot = TRUE,
  base_size = 14
)
}
\arguments{
\item{x}{An object of class \code{theil_list}, as created by calling \code{\link[surveil]{theil}} on a list of fitted \code{surveil} models.}

\item{fill}{Fill color for credible intervals}

\item{alpha}{Transparency of credible interval fill color}

\item{scale}{Multiply Theil's T by \code{scale} for readability}

\item{plot}{If \code{FALSE}, return a list of \code{ggplot}s.}

\item{base_size}{Passed to \code{theme_classic} to control size of plot elements (e.g., text)}

\item{color}{Line color}
}
\value{
If \code{plot = FALSE}, returns a list of \code{ggplot}s; otherwise, the \code{ggplots} will be drawn to the current plotting device using \code{\link[gridExtra]{grid.arrange}}.
}
\description{
Plot the between area, within area, and total (between + within) inequality as measured by Theil's index.
}
