% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.theil}
\alias{print.theil}
\title{print Theil's index}
\usage{
\method{print}{theil}(x, scale = 100)
}
\arguments{
\item{rate_scale}{Print rates and rate differences as per \code{rate_scale}.}
}
\description{
print Theil's index
}
