% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard}
\alias{standard}
\title{2000 U.S. standard million population}
\format{
An object of class \code{data.frame} with 19 rows and 3 columns.
}
\source{
National Cancer Institute. Standard Populations - 19 Age Groups. Accessed at \url{https://seer.cancer.gov/stdpopulations/stdpop.19ages.html} on Oct. 8, 2021.
}
\usage{
standard
}
\description{
2000 U.S. standard million population
}
\examples{
data(standard)
head(standard)
}
\keyword{datasets}
