% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fit}
\alias{fit}
\alias{plot.surveil}
\alias{print.surveil}
\alias{bridge_sampler.surveil}
\alias{waic.surveil}
\alias{loo.surveil}
\title{Plot disease risk time trends}
\source{
Watanabe, S. (2010). Asymptotic equivalence of Bayes cross validation and widely application information criterion in singular learning theory. Journal of Machine Learning Research 11, 3571-3594.

Vehtari, A., Gelman, A., and Gabry, J. (2017a). Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC. Statistics and Computing. 27(5), 1413--1432. doi:10.1007/s11222-016-9696-4 (journal version, preprint arXiv:1507.04544).

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry, J. (2019). Pareto smoothed importance sampling. preprint arXiv:1507.02646

Vehtari, A., Gelman, A., and Gabry, J. (2017a). Practical Bayesian model evaluation using leave-one-out cross-validation and WAIC. Statistics and Computing. 27(5), 1413--1432. \link{doi:10.1007/s11222-016-9696-4} (journal version, preprint \link{arXiv:1507.04544}).

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry, J. (2019). Pareto smoothed importance sampling. preprint \link{arXiv:1507.02646}
}
\usage{
\method{plot}{surveil}(fit, rate_scale = 1, base_size = 14)

\method{print}{surveil}(fit, ...)

\method{bridge_sampler}{surveil}(fit, ...)

\method{waic}{surveil}(fit)

\method{loo}{surveil}(fit, save_psis = FALSE, cores = getOption("mc.cores", 1))
}
\arguments{
\item{fit}{A fitted \code{surveil} model.}

\item{base_size}{Passed to \code{theme_classic()} to control size of plot components (text).}

\item{...}{Other arguments passed to \code{\link[bridgesampling]{bridge_sampler}}.}
}
\value{
a \code{ggplot} object
}
\description{
Plotting method for \code{surveil} model results

Print a summary of \code{surveil} model results to the console

Computes log marginal likelihood via bridge sampling. This is needed to calculate the Bayes factor for model comparison using the \code{\link{bridgesampling}} package.

Calculates WAIC for a fitted \code{surveil} model using \code{\link[loo]{waic}}.

Implements \code{\link[loo]{loo.array}} for \code{surveil} models.
}
\details{
Please see \code{\link[bridgesampling]{bridge_sampler}} for details and references.
}
\examples{

data(msa)

library(tidyverse)
   data(msa)
   dfw <- msa[grep("Dallas", msa$MSA), ]
   fit <- stan_rw(dfw)

  ## plot time-varying risk with 95\\% credible intervals
 plot(fit)

 ## as a ggplot, you can customize the output
plot(fit) +
 scale_x_continuous(breaks = 1:19, 
                    labels = 1999:2017,
                    name = NULL
                    )

}
\seealso{
\code{\link[surveil]{stan_rw}}

\code{\link[surveil]{stan_rw}}

\code{\link[bridgesampling]{bayes_factor}}  \code{\link[bridgesampling]{post_prob}} \code{\link[surveil]{stan_rw}}

\code{\link[surveil]{stan_rw}}

\code{\link[surveil]{stan_rw}}
}
\author{
Connor Donegan (Connor.Donegan@UTSouthwestern.edu)
}
