% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{surveil_diff}
\alias{surveil_diff}
\alias{plot.surveil_diff}
\alias{print.surveil_diff}
\title{Plot surveil_diff objects for analyses of inequality}
\usage{
\method{plot}{surveil_diff}(
  x,
  col = "black",
  fill = "gray80",
  plot = TRUE,
  rate_scale = 1e+05,
  PAR = TRUE,
  base_size = 14
)

\method{print}{surveil_diff}(x, rate_scale = 1)
}
\arguments{
\item{col}{Line color}

\item{fill}{Fill color for credible intervals, passed to \code{geom_ribbon}}

\item{plot}{If \code{plot = FALSE}, a list of \code{ggplot}s will be returned}

\item{rate_scale}{Print rates and rate differences as per \code{rate_scale}.}

\item{PAR}{Return population attributable risk? IF \code{FALSE}, then the rate ratio will be used instead of PAR.}

\item{base_size}{Passed to \code{theme_classic} to control size of plot elements (e.g., text)}
}
\description{
Plot surveil_diff objects for analyses of inequality

print surveil_diff objects for analyses of inequality
}
