% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{theil.surveil}
\alias{theil.surveil}
\alias{theil.list}
\alias{theil}
\alias{theil2}
\alias{theil.data.frame}
\title{Theil's inequality index}
\source{
Conceicao, P. and P. Ferreira (2000). The young person's guide to the Theil Index: Suggesting intuitive interpretations and exploring analytical applications. University of Texas Inequality Project. UTIP Working Paper Number 14. Accessed May 1, 2021 from \url{https://utip.gov.utexas.edu/papers.html}

Theil, Henri (1972). \emph{Statistical Decomposition Analysis.} Amsterdan, The Netherlands and London, UK: North-Holland Publishing Company.

Shannon, Claude E. and Weaver, Warren (1963). \emph{The Mathematical Theory of Communication}. Urbana and Chicago, USA: University if Illinois Press.
}
\usage{
\method{theil}{surveil}(x)

\method{theil}{list}(x)

theil(x, group, time, detail = FALSE)

theil2(Count, Population, rates, total = TRUE)

\method{theil}{data.frame}(x, detail = FALSE)
}
\arguments{
\item{x}{A data.frame with columns "Count" (for disease case count or deaths) and "Population" (for population at risk), and optional columns for "time" and "geog" (geographic grouping variable, such as state or city name).

If \code{x} is a \code{data.frame} and it has a geographic grouping variable (as indicated by a column named "geog"), a between-geographies component of inequality will be added to the index, such that Theil's T will be the sum of within-geography inequality and between-georgaphy inequality. Providing a \code{geog} variable that contains only one value (a single georgraphic area) is the same as not providing any \code{geog} variable, such that Theil's T indexes the within-geography inequality only.

If \code{x} is a \code{data.frame} and it has a column named "time", then Theil's index will be calculated separately for each time period.}

\item{detail}{If \code{detail = TRUE}, the return value will be a list with each component of Theil's T. Theil'T is the sum of all its parts.}
}
\value{
A named list with the following elements: \code{summary}, A \code{data.frame} summarizing the posterior probability distribution for Theil's T, including the mean and 95\\% credible interval for each time period; and \code{samples}, A \code{data.frame} with MCMC samples for Theil's T.

The results depend on the input. For a non-nested data structure, a numeric value is returned for each time period of observation.
}
\description{
Calculates Theil's entropy-based measure of inequality for a given set of disease incidence rates and populations at risk. For nested data structures, within-group (\code{W}) and between-group (\code{B}) components of inequality are calculated, as well as total inequality (\code{T = W + B}).
}
\details{
Theil's index is a good index of inequality in disease and mortality burdens when multiple groups are being considered, as is typical of geospatial analysis. It provides a summary measure of inequality across a set of demographic groups that may be tracked over time. Also, it is interesting because it is additive, and thus admits of simple decompositions.

The index measures discrepancies between a population's share of the disease burden, \code{omega}, and their share of the population, \code{eta}. A situation of zero inequality would imply that each population's share of cases is equal to its population share, or, \code{omega=eta}. Each population's contribution to total inequality is calculated as:\preformatted{             T_i = omega_i * [log(omega_i/eta_i)],
}

the log-ratio of case-share to population-share, weighted by their share of cases. Theil's index for all areas is the sum of each area's T_i:\preformatted{             T = sum_(i=1)^n T_i.
}

Theil's T is thus a weighted mean of log-ratios of case shares to population shares, where each log-ratio (which we may describe as a raw inequality score) is weighted by its share of total cases. The index has a minimum of zero and a maximum of \code{log(N)}, where \code{N} is the number of units (e.g., number of counties).
Theil's index is based on Shannon's information theory, he used it to study a variety of topics, including income inequality and racial segregation. Theil's index is often of great interest because it is additive across multiple scales, such as when the data has a nested structure to it (e.g., demographic groups within states). The Texas Inequality Project provides introductions to, and examples of using, the Theil index (Conceicao and Ferreira, 2000). However, this \code{R} function is just a simple implementation for `flat' or non-nested data structures (e.g., a set of counties).
}
\examples{

Count <- c(10, 12, 3, 111)
Pop <- c(1000, 1200, 4000, 9000)
theil2(Count, Pop)
}
\seealso{
\code{\link[surveil]{theil2}}

\code{\link[surveil]{theil}}
}
