% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apc.R
\name{print.apc_ls}
\alias{print.apc_ls}
\alias{plot.apc_ls}
\title{Methods for APC objects}
\usage{
\method{print}{apc_ls}(x, digits = 1, max = 10, ...)

\method{plot}{apc_ls}(
  x,
  cumulative = FALSE,
  style = c("mean_qi", "lines"),
  M = 250,
  col = "black",
  fill = "black",
  alpha = ifelse(style == "mean_qi", 0.5, 0.7),
  lwd = ifelse(style == "mean_qi", 1, 0.05),
  base_size = 14,
  ...
)
}
\arguments{
\item{x}{An \code{apc_ls} object returned by \code{\link[surveil]{apc}}}

\item{digits}{Print this many digits (passed to \code{\link[base]{print.data.frame}})}

\item{max}{Maximum number of time periods (rows) to print}

\item{...}{Additional arguments}

\item{cumulative}{Plot cumulative percent change? Defaults to \code{cumulative = FALSE}}

\item{style}{If \code{style = "mean_qi"}, then the posterior mean and 95 percent credible interval will be plotted; if \code{style = "lines"}, then \code{M} samples from the joint probability distribution of the annual rates will be plotted.}

\item{M}{If \code{style = "lines"}, then \code{M} is the number of samples from the posterior distribution that will be plotted; the default is \code{M = 250}.}

\item{col}{Line color}

\item{fill}{Fill color for the 95 percent credible interval}

\item{alpha}{For \code{style = "mean_qi"}, this controls the transparency for the credible interval (passed to \code{\link[ggplot2]{geom_ribbon}}) and defaults to \code{alpha = 0.5}; for \code{style = "lines"}, this controls the transparency of the lines and defaults to \code{alpha = 0.7}.}

\item{lwd}{Line width}

\item{base_size}{Size of plot attributes, passed to `\code{\link[ggplot2]{theme_classic}}}
}
\description{
Methods for APC objects
}
\details{
\subsection{print.apc_ls}{

Any additional arguments (\code{...}) to the print method will be  passed to \code{\link[base]{print.data.frame}}
}

\subsection{plot.apc_ls}{

Any additional arguments (\code{...}) will be passed to \code{\link[ggplot2]{theme}}.
}
}
