% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-diff.R
\name{group_diff}
\alias{group_diff}
\title{Measures of pairwise inequality}
\source{
Menvielle, G, Kulhanaova, I, Machenbach, JP. Assessing the impact of a public health intervention to reduce social inequalities in cancer. In: Vaccarella, S, Lortet-Tieulent, J, Saracci, R, Conway, D, Straif, K, Wild, CP, editors. Reducing Social Inequalities in Cancer: Evidence and Priorities for Research. Geneva, Switzerland: WHO Press, 2017:185-192.
}
\usage{
group_diff(fit, target, reference, samples = FALSE)
}
\arguments{
\item{fit}{A fitted \code{surveil} time series model}

\item{target}{The name (character string) of the disadvantaged group that is the target of inference.}

\item{reference}{The name (character string) of the reference group to which \code{target} will be compared.}

\item{samples}{If \code{TRUE}, return MCMC samples for the inequality measures. The default is \code{FALSE}.}
}
\value{
A list, also of class "surveil_diff", with the following elements: \describe{
\item{summary}{A tibble with a summary of posterior distributions (mean and 95 percent cred. intervals) for the target group incidence rate, the RD, RR, PAR, and excess cases.}
\item{cumulative_cases}{Summary of the posterior distribution for the cumulative number of excess cases and the PAR (mean and 95 percent cred. intervals)}
\item{groups}{Character string with target and reference population names}
}
If \code{samples = TRUE}, then, in addition to the above items, a list called \code{samples} is returned; it contains:
\describe{
\item{annual}{A data frame of MCMC samples for each quantity of interest (target and reference rates, RD, RR, PAR, and EC, as well as \code{Trend_Cases = Rate * Population}. Indexed by time.}
\item{cumulative_cases}{MCMC samples of the cumulative number of excess cases.}
}
}
\description{
Calculate pairwise measures of health inequality from a fitted \code{surveil} time series model, with credible intervals and MCMC samples. Calculates absolute and fractional rate differences (RD and population attributable risk (PAR)), rate ratios, and excess cases.
}
\details{
For the following calculations, the terms \code{reference} and \code{target} refer to incidence rates for the respective groups; \code{p} is the size of the target population. (Target is the group that is the 'target' of our inferences, so that it is the numerator in rate ratios, etc.) The following measures are calculated by \code{group_diff}:\preformatted{# rate difference
RD = target - reference
# population attributable fraction
PAR = RD/target = (RR - 1)/RR
# rate ratio
RR = target/reference
# excess cases
EC = RD * p
}

As the math concisely communicates, the PAR is defined as the fraction of risk in the target population that would have been removed had the target rate equaled the reference rate (Menvielle et al. 2017).
}
\seealso{
\code{\link[surveil]{plot.surveil_diff}} \code{\link[surveil]{print.surveil_diff}} \code{\link[surveil]{theil}}
}
\author{
Connor Donegan (Connor.Donegan@UTSouthwestern.edu)
}
