% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\alias{student_t}
\alias{lkj}
\title{Prior distributions}
\source{
Stan Development Team. Stan Functions Reference Version 2.27. \url{https://mc-stan.org/docs/2_27/functions-reference/lkj-correlation.html}
}
\usage{
student_t(df = 20, location = 0, scale)

lkj(eta)
}
\arguments{
\item{df}{Degrees of freedom, positive numeric value(s)}

\item{location}{Location parameter}

\item{scale}{Scale parameter}

\item{eta}{The shape parameter for the LKJ prior}
}
\value{
An object of class \code{prior} which will be used internally by \strong{surveil} to set parameters of prior distributions.
}
\description{
Prior distributions
}
\details{
The prior distribution functions are used to set the values of prior parameters.

Users can control the values of the parameters, but the distribution (model) itself is fixed. The first log-rate (eta a\link{t}, t=1) and the scale parameters (sigma) are assigned Student's t prior distribution. (The scale parameter, sigma, is constrained to be positive, making it a half-Student's t prior.) For correlated time series, the correlation matrix is assigned the LKJ prior.
\subsection{Parameterizations}{

For details on how any distribution is parameterized, see the Stan Language Functions Reference document: \url{https://mc-stan.org/users/documentation/}.
}

\subsection{LKJ prior}{

The LKJ prior for correlation matrix has a single parameter, eta (eta > 0). If \code{eta=1}, then you are placing a uniform prior on any K-by-K correlation matrix. For eta > 1, there is a higher probability on the identity matrix, such that as eta increases beyond 1, you are expressing greater scepticism towards large correlations. If 0 < eta < 1, then you will be expressing scepticism towards correlations of zero and favoring non-zero correlations. See Stan documentation: \url{https://mc-stan.org/docs/2_27/functions-reference/lkj-correlation.html}.
}
}
\examples{

prior <- list()
prior$eta_1 <- student_t(df = c(20, 20, 20), location = c(-5, -5, -5), scale = c(3, 3, 3))
# By default, all `df = 20`
prior$sigma <- student_t(location = c(0,0,0), scale = c(1,1,1))
prior$omega <- lkj(2)
\dontrun{
dfw <- dplyr::filter(msa, grepl("Dallas", MSA))
fit <- stan_rw(dfw, time = Year, group = Race, prior = prior)
plot(fit)
}

}
