% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.stand_surveil}
\alias{print.stand_surveil}
\alias{plot.stand_surveil}
\title{Methods for standardized rates}
\usage{
\method{print}{stand_surveil}(x, scale = 1, digits = 3, ...)

\method{plot}{stand_surveil}(x, scale = 1, base_size = 14, col = "black", fill = "gray80", ...)
}
\arguments{
\item{x}{An object of \code{stand_surveil} obtained by calling \code{\link[surveil]{standardize}} on a fitted \code{surveil} model}

\item{scale}{Scale the rates by this amount; e.g., \code{scale = 100e3} will print rates per 100,000 at risk.}

\item{digits}{Number of digits to print}

\item{...}{additional arguments}

\item{base_size}{Passed to \code{theme_classic()} to control size of plot components (text).}

\item{col}{Line color}

\item{fill}{Fill color for the 95 percent credible intervals}
}
\description{
Print and plot methods for \code{stand_surveil} (standardized rates obtained from a fitted \code{surveil} model)
}
\details{
\subsection{print.stand_surveil}{

Any additional arguments (\code{...}) will be  passed to \code{\link[base]{print.data.frame}}
}

\subsection{plot.stand_surveil}{

Any additional arguments (\code{...}) will be passed to `\code{\link[ggplot2]{theme}}.
}
}
\seealso{
\code{\link[surveil]{standardize}} \code{\link[surveil]{stan_rw}}
}
