% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveil-package.R
\docType{package}
\name{surveil-package}
\alias{surveil-package}
\alias{surveil}
\title{The 'surveil' package}
\description{
Public health surveillance with time series models for disease incidence and mortality, with methods for inference on measures of health inequality, cumulative and period percent change, and age-standardized rates. An accessible interface to Stan, a state-of-the-art platform for Bayesian inference.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
