% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.surveil}
\alias{plot.surveil}
\alias{print.surveil}
\title{Plot disease risk time trends}
\usage{
\method{plot}{surveil}(x, scale = 1, base_size = 14, ...)

\method{print}{surveil}(x, scale = 1, ...)
}
\arguments{
\item{x}{A fitted \code{surveil} model.}

\item{scale}{Scale the rates by this amount; e.g., \code{scale = 100e3} will print rates per 100,000 at risk.}

\item{base_size}{Passed to \code{theme_classic()} to control size of plot components (text).}

\item{...}{additional arguments passed to \code{\link[base]{print.data.frame}}}
}
\value{
a \code{ggplot} object
}
\description{
Plotting method for \code{surveil} model results

Print a summary of \code{surveil} model results to the console
}
\examples{

\dontrun{
library(ggplot2)
   data(msa)
   dfw <- msa[grep("Dallas", msa$MSA), ]
   fit <- stan_rw(dfw, time = Year, group = Race)

  ## plot time-varying risk with 95\\% credible intervals
 plot(fit)
 plot(fit, legend.text = element_text(size = 12))

 ## as a ggplot, you can customize the output
plot(fit) +
 scale_x_continuous(breaks = 1:19, 
                    labels = 1999:2017,
                    name = NULL
                    )
}
}
\seealso{
\code{\link[surveil]{stan_rw}}

\code{\link[surveil]{stan_rw}} \code{\link[surveil]{plot.surveil}}
}
\author{
Connor Donegan (Connor.Donegan@UTSouthwestern.edu)
}
