% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.surveil}
\alias{plot.surveil}
\alias{print.surveil}
\title{Methods for fitted \code{surveil} models}
\usage{
\method{plot}{surveil}(
  x,
  scale = 1,
  style = c("mean_qi", "lines"),
  facet = FALSE,
  base_size = 14,
  palette = "Dark2",
  M = 250,
  alpha,
  lwd,
  ...
)

\method{print}{surveil}(x, scale = 1, ...)
}
\arguments{
\item{x}{A fitted \code{surveil} model.}

\item{scale}{Scale the rates by this amount; e.g., \code{scale = 100e3} will print rates per 100,000 at risk.}

\item{style}{If \code{style = "mean_qi"}, then the posterior mean and 95 percent credible interval will be plotted; if \code{style = "lines"}, then \code{M} samples from the joint probability distribution of the annual rates will be plotted.}

\item{facet}{If \code{facet = TRUE}, \code{\link[ggplot2]{facet_wrap}} will be used instead of differentiating by line color.}

\item{base_size}{Passed to \code{theme_classic()} to control size of plot components (text).}

\item{palette}{For multiple groups, choose the color palette. For a list of options, see \code{\link[ggplot2]{scale_color_brewer}}. The default is \code{palette = "Dark2"}. Not used if \code{facet = TRUE}.}

\item{M}{If \code{style = "lines"}, then \code{M} is the number of samples from the posterior distribution that will be plotted; the default is \code{M = 250}.}

\item{alpha}{For \code{style = "lines"}; numeric value from zero to one indicating transparency of lines; passed to \code{\link[ggplot2]{geom_line}}. For `style = "mean_qi", this controls the transparency of the shaded credible interval; passed to \code{\link[ggplot2]{geom_ribbon}}.}

\item{lwd}{For \code{style = "lines"}; numeric value indicating linewidth. Passed to \code{\link[ggplot2]{geom_line}}}

\item{...}{additional arguments passed to \code{\link[base]{print.data.frame}}}
}
\value{
a \code{ggplot} object
}
\description{
Plotting method for \code{surveil} model results

Print a summary of \code{surveil} model results to the console
}
\examples{

\dontrun{
library(ggplot2)
   data(msa)
   dfw <- msa[grep("Dallas", msa$MSA), ]
   fit <- stan_rw(dfw, time = Year, group = Race)

  ## plot time-varying risk with 95\\% credible intervals
 plot(fit)
 plot(fit, legend.text = element_text(size = 12))

 ## as a ggplot, you can customize the output
plot(fit) +
 scale_x_continuous(breaks = 1:19, 
                    labels = 1999:2017,
                    name = NULL
                    )
}
}
\seealso{
\code{\link[surveil]{stan_rw}}

\code{\link[surveil]{stan_rw}} \code{\link[surveil]{plot.surveil}}
}
\author{
Connor Donegan (Connor.Donegan@UTSouthwestern.edu)
}
