% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-diff.R
\name{surveil_diff}
\alias{surveil_diff}
\alias{plot.surveil_diff}
\alias{print.surveil_diff}
\title{Methods for \code{surveil_diff} objects}
\usage{
\method{plot}{surveil_diff}(
  x,
  col = "black",
  fill = "gray80",
  plot = TRUE,
  scale = 1e+05,
  PAR = TRUE,
  base_size = 14,
  ...
)

\method{print}{surveil_diff}(x, scale = 1, ...)
}
\arguments{
\item{x}{Object of class \code{surveil_diff}, as returned by calling \code{group_diff} on a fitted \code{surveil} model}

\item{col}{Line color}

\item{fill}{Fill color for credible intervals, passed to \code{geom_ribbon}}

\item{plot}{If \code{plot = FALSE}, a list of \code{ggplot}s will be returned}

\item{scale}{Print rates and rate differences as per \code{scale} at risk, e.g., per 10,000 at risk.}

\item{PAR}{Return population attributable risk? IF \code{FALSE}, then the rate ratio will be used instead of PAR.}

\item{base_size}{Passed to \code{theme_classic} to control size of plot elements (e.g., text)}

\item{...}{additional print arguments}
}
\description{
Methods for \code{surveil_diff} objects

print surveil_diff objects for analyses of inequality
}
