% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil.R
\name{plot.theil}
\alias{plot.theil}
\alias{plot.theil_list}
\alias{print.theil}
\alias{print.theil_list}
\title{Methods for Theil's index}
\usage{
\method{plot}{theil}(
  x,
  col = "black",
  fill = "gray80",
  alpha = 0.5,
  base_size = 14,
  scale = 100,
  labels = x$summary$time,
  ...
)

\method{plot}{theil_list}(
  x,
  col = "black",
  fill = "black",
  alpha = 0.25,
  between_title = "Between",
  within_title = "Within",
  total_title = "Total",
  scale = 100,
  plot = TRUE,
  base_size = 14,
  ...
)

\method{print}{theil}(x, scale = 100, digits = 3, ...)

\method{print}{theil_list}(x, scale = 100, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \verb{thiel' or }theil_list\verb{, as returned by calling }theil\verb{on a list of fitted}surveil` models}

\item{col}{Line color}

\item{fill}{Fill color}

\item{alpha}{Transparency of credible interval fill color}

\item{base_size}{Passed to \code{theme_classic} to control size of plot elements (e.g., text)}

\item{scale}{Scale Theil's index by \code{scale}}

\item{labels}{x-axis labels (time periods)}

\item{...}{additional arguments}

\item{between_title}{Plot title for the between geography component of Theil's T; defaults to "Between".}

\item{within_title}{Plot title for the within geography component of Theil's T; defaults to "Within".}

\item{total_title}{Plot title for Theil's index; defaults to "Total".}

\item{plot}{If \code{FALSE}, return a list of \code{ggplot}s.}

\item{digits}{number of digits to print (passed to \code{\link[base]{print.data.frame}})}
}
\description{
Printing and plotting methods for Theil's inequality index
}
\examples{
 \dontrun{
 dfw <- msa[grep("Dallas", msa$MSA), ]
 fit <- stan_rw(dfw, time = Year, group = Race)
 theil.dfw <- theil(fit)
 plot(theil.dfw)
}
}
\seealso{
\code{\link[surveil]{theil}}
}
